Add option to include version in package name to make versions coinstallable
Building some crates requires installing more than one version of a
library simultaneously to satisfy various versioned dependencies in the
dependency tree.  Add an option --multi to include the ABI version in
the package name, producing co-installable packages for multiple
versions.

The ABI version consists of the major number if non-zero, or 0.minor
with a major number of 0.  The version number included in the package
name always includes the '.' (which cannot appear in crate names), to
avoid any possible conflict between the versioned name of a crate and
the unversioned name of another crate.

Packages with versioned names provide and conflict with the same version
of the corresponding unversioned package.