Always include ABI version in package name; drop --multi
Using versioned Provides to allow packages to include or omit the
package name doesn't work in all cases; in particular, in some cases, it
could allow satisfying the Debian package dependencies without
satisfying the cargo dependencies, by installing multiple different
providers of the package name simultaneously, none of which meet all the
constraints.

Switch to always including versions in the package names, and change
dependencies to include the version of the target package.  Add
error-handling to dependency processing, to bail out on unrepresentable
dependencies.