Change functionality of copyright_fromgit function
Instead of returning all commiters list with years return only first and
latest contribution year from git. This is done because it does not feel
to put every one who commited in git repo as copyright holder for
project. So only extracting contribution year seems fine. (Though
technically it is also not correct as authorship should be claimed by
project author and not package maintainers).