Handle optional overrides gracefully
Some time when optional override were present but the specific override
we are looking for (eg. missing Files overrides) were not present then
default was not executed as it was in else block of checking if
overrides were present.

To fix this we implement std::default::Default trait for Overrides and
provide a default with all None fields. During beginning of
debian_copyright we check if overrides were present or not, if not we
use default overrides there by making it like there were no override at
all.

For implementing this we also derived Clone as we had to be able to
create owned copy of overrides passed by do_package.