handle namespaced features
cargo will now inject a feature named after an optional dependency (and
use it on the RHS of the feature definition) if needed, so we don't need
to do it manually here. if no such feature exists it means that all
references to the optional dependency are namespaced with 'dep:' prefix,
and that the dependency itself should not(!) be exposed as feature.

any features depending on such a prefixed optional dependency will still
caused the required `Depends:` to be generated, we just skip the
corresponding `Provides:` for the no-longer-usable pseudo-feature.

tested various combinations using the local crate support, they all seem
to do the right thing. it might be prudent to write test cases for all
combinations we expect to handle.

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>