tests: fix once-cell 1.16
it renames an optional dependency that is also re-exposed under the
original name, and the only difference is '_' vs '-' which debcargo
normalization of crate/feature names doesn't like.

since the renamed optional name is just for internal usage (documented
as such) we can use a different one that doesn't clash.

Signed-off-by: Fabian Grünbichler <debian@fabian.gruenbichler.email>